\name{bibentry_key}
\alias{bibentry_key}
\title{Give the key associated with a bibentry element}
\description{Give the key associated with a bibentry element}
\usage{
bibentry_key(x)
}
\arguments{
  \item{x}{a single bibentry element.}
}
\details{
  This is a convenience function to get the "key" attribute of a
  reference represented by a bibentry element.

  There should be a better way to do this, I must be missing something
  here.
}
\value{
  a character string
}
\author{Georgi N. Boshnakov}

\keyword{ bibtex }
\keyword{programming}
