\name{Rdo_locate_predefined_section}
\alias{Rdo_locate_predefined_section}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Locate the position of a predefined section}
\description{Locate the position of a predefined section.}
\usage{
Rdo_locate_predefined_section(rdo, sec)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{rdo}{an Rd object.}
  \item{sec}{the required ection, a character string.}
}
\details{
  This returns the position of a top level section in an Rd object.

  todo: more care needed.
}
\value{
  a numeric vector
}
% \references{
% %% ~put references to the literature/web site here ~
% }
\author{Georgi N. Boshnakov}
% \note{
% %%  ~~further notes~~
% }
%
% %% ~Make other sections like Warning with \section{Warning }{....} ~
%
% \seealso{
% %% ~~objects to See Also as \code{\link{help}}, ~~~
% }
% \examples{
% }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{RdoProgramming}

