\name{update_aliases_tmp}

\alias{update_aliases_tmp}

%- Also NEED an '\alias' for EACH other topic documented here.

\title{Update aliases for methods in Rd objects}

\description{Update aliases for methods in Rd objects}

\usage{
update_aliases_tmp(rdo, package = NULL)
}

%- maybe also 'usage' for other objects documented here.

\arguments{
  \item{rdo}{an Rd object}
  \item{package}{the name of a package, a character string.}
}

\details{
  This is a quick fix. todo: complete it!
}

\value{
  the updated Rd object
}



\author{Georgi N. Boshnakov}




%% ~Make other sections like Warning with \section{Warning }{....} ~






% Add one or more standard keywords, see file 'KEYWORDS' in the

% R documentation directory.

\keyword{RdoS4}

