\name{inspect_slots}

\alias{inspect_slots}

%- Also NEED an '\alias' for EACH other topic documented here.

\title{
  Inspect the slots of an S4 class
}

\description{
  Inspect the slots of an S4 class.
}
\usage{
inspect_slots(rdo, final = TRUE)
}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{rdo}{an Rd object.}
  \item{final}{if not TRUE insert text with suggestions, otherwise
    comment the suggestions out.
  }
}
\author{Georgi N. Boshnakov}
%% ~Make other sections like Warning with \section{Warning }{....} ~
\keyword{RdoS4}

